% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{read_interpretation}
\alias{read_interpretation}
\title{Read an interpretation from a JSON file}
\usage{
read_interpretation(path)
}
\arguments{
\item{path}{Path to an interpretation JSON file.}
}
\value{
An \code{interpretation} object.
}
\description{
Read an interpretation from a JSON file
}
