% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\docType{data}
\name{RENDER_TEMPLATE_VERSION}
\alias{RENDER_TEMPLATE_VERSION}
\title{Render template version}
\format{
An object of class \code{character} of length 1.
}
\usage{
RENDER_TEMPLATE_VERSION
}
\description{
Bumped whenever \code{\link[=render_paragraph]{render_paragraph()}}'s template changes; recorded in the
run-level manifest (\code{\link[=build_synthesis_manifest]{build_synthesis_manifest()}}).
}
\keyword{datasets}
