% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation.R
\name{write_interpretation}
\alias{write_interpretation}
\title{Write an interpretation to a JSON file}
\usage{
write_interpretation(interp, path)
}
\arguments{
\item{interp}{An \code{interpretation} object.}

\item{path}{Output file path.}
}
\value{
\code{path}, invisibly.
}
\description{
Write an interpretation to a JSON file
}
